fx: move DinoStomp
