+++ meson.build
@@ -92,6 +92,7 @@ sources = [
   'src/game/draw.c',
   'src/game/effect_routines/bubbles.c',
   'src/game/effect_routines/chain_block.c',
+  'src/game/effect_routines/dino_stomp.c',
   'src/game/effects/blood.c',
   'src/game/effects/exploding_death.c',
   'src/game/fmv.c',
@@ -146,7 +147,6 @@ sources = [
   'src/game/objects/effects/blood.c',
   'src/game/objects/effects/body_part.c',
   'src/game/objects/effects/bubble.c',
-  'src/game/objects/effects/dino_stomp.c',
   'src/game/objects/effects/earthquake.c',
   'src/game/objects/effects/explosion.c',
   'src/game/objects/effects/finish_level.c',
