+++ src/game/effect_routines/dino_stomp.h
@@ -2,4 +2,4 @@
 
 #include "global/types.h"
 
-void DinoStomp(ITEM_INFO *item);
+void FX_DinoStomp(ITEM_INFO *item);
