+++ src/global/vars.c
@@ -1,6 +1,6 @@
 #include "game/effect_routines/bubbles.h"
 #include "game/effect_routines/chain_block.h"
-#include "game/objects/effects/dino_stomp.h"
+#include "game/effect_routines/dino_stomp.h"
 #include "game/objects/effects/earthquake.h"
 #include "game/objects/effects/explosion.h"
 #include "game/objects/effects/finish_level.h"
@@ -19,9 +19,9 @@ char *GameMemoryPointer = NULL;
 int32_t g_FPSCounter = 0;
 
 void (*g_EffectRoutines[])(ITEM_INFO *item) = {
-    Turn180,       DinoStomp, LaraNormal,    FX_Bubbles,   FinishLevel,
-    EarthQuake,    Flood,     RaisingBlock,  Stairs2Slope, DropSand,
-    PowerUp,       Explosion, LaraHandsFree, FxFlipMap,    LaraDrawRightGun,
+    Turn180,       FX_DinoStomp, LaraNormal,    FX_Bubbles,   FinishLevel,
+    EarthQuake,    Flood,        RaisingBlock,  Stairs2Slope, DropSand,
+    PowerUp,       Explosion,    LaraHandsFree, FxFlipMap,    LaraDrawRightGun,
     FX_ChainBlock, Flicker,
 };
